basic design of client/server communication
as of now, it allows for the client and server to send one-way communications back and forth.

ideally, there should be transmissions every x seconds to share data or updates if anything.

multiple jobs can go to one worker ... what matters is that the master is able to appropriately send the right response back to the right client